# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Shopping Agent API
  version: 0.1.0
  description: 智能识物比价购物 AI APP 助手 API 契约。
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /search-tasks:
    post:
      tags: [SearchTasks]
      summary: 创建搜索任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSearchTaskRequest'
      responses:
        '200':
          description: 搜索任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTaskResponse'
    get:
      tags: [SearchTasks]
      summary: 查询搜索历史
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 搜索历史
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTaskSummaryPageResponse'
  /search-tasks/{searchTaskId}:
    get:
      tags: [SearchTasks]
      summary: 查询搜索任务详情
      parameters:
        - $ref: '#/components/parameters/SearchTaskId'
      responses:
        '200':
          description: 搜索任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTaskResponse'
  /search-tasks/{searchTaskId}/refine:
    post:
      tags: [SearchTasks]
      summary: 自然语言追加筛选并刷新结果
      parameters:
        - $ref: '#/components/parameters/SearchTaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefineSearchTaskRequest'
      responses:
        '200':
          description: 追加筛选结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefineSearchTaskResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    SearchTaskId:
      name: searchTaskId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
    SourceType:
      type: string
      enum: [mock, official_api, sample_dataset]
    Platform:
      type: string
      enum: [jd, taobao, pdd, tmall, other]
    SortMode:
      type: string
      enum: [comprehensive, price_asc, sales_desc, rating_desc]
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          example: '199.00'
        currency:
          type: string
          example: CNY
    SuggestionCard:
      type: object
      required: [cardId, type, title, payload, priority]
      properties:
        cardId: { type: string }
        type: { type: string, enum: [filter, sort, price_history, similar, official] }
        title: { type: string }
        payload:
          type: object
          additionalProperties: true
        priority: { type: integer }
    PlatformStats:
      type: object
      required: [platform, lowestPrice, averagePrice, productCount]
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        lowestPrice:
          $ref: '#/components/schemas/Money'
        averagePrice:
          $ref: '#/components/schemas/Money'
        productCount: { type: integer }
    ApiMeta:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
...
    Recognition:
      type: object
      required: [recognitionId, imageId, category, keywords, confidence, status, createdAt]
      properties:
        recognitionId: { type: integer, format: int64 }
        imageId: { type: integer, format: int64 }
        category: { type: string }
        brand: { type: string, nullable: true }
        model: { type: string, nullable: true }
        keywords:
          type: array
          items: { type: string }
        attributes:
          type: object
          additionalProperties: true
        suggestionCards:
          type: array
          items:
            $ref: '#/components/schemas/SuggestionCard'
        confidence: { type: number, format: double }
        aiProvider:
          type: string
          example: mock
        fallbackUsed:
          type: boolean
          description: Ark/VLM/LLM 不可用时是否已回退到 mock 或规则解析。
        explanation:
          type: string
          nullable: true
        notices:
          type: array
          items: { type: string }
        status: { type: string, enum: [pending, succeeded, failed] }
        createdAt: { type: string, format: date-time }
...
    CreateSearchTaskRequest:
      type: object
      properties:
        recognitionId: { type: integer, format: int64, nullable: true }
        query: { type: string, nullable: true }
        platforms:
          type: array
          items:
            $ref: '#/components/schemas/Platform'
        sourceType:
          $ref: '#/components/schemas/SourceType'
        filters:
          type: object
          additionalProperties: true
        sortBy:
          $ref: '#/components/schemas/SortMode'
    SearchTaskItem:
      type: object
      required: [platformProductId, productId, platform, title, price, matchScore, matchReasons, sourceType, tags, salesVolume, rating, isOfficial, isSelfOperated]
      properties:
        platformProductId: { type: integer, format: int64 }
        productId: { type: integer, format: int64 }
        platform:
          $ref: '#/components/schemas/Platform'
        title: { type: string }
        imageUrl: { type: string, nullable: true }
        price:
          $ref: '#/components/schemas/Money'
        originalPrice:
          $ref: '#/components/schemas/Money'
        url: { type: string, nullable: true }
        shopName: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        salesVolume: { type: integer }
        rating: { type: number, format: double }
        isOfficial: { type: boolean }
        isSelfOperated: { type: boolean }
        matchScore: { type: number, format: double }
        matchReasons:
          type: array
          items: { type: string }
        sourceType:
          $ref: '#/components/schemas/SourceType'
        updatedAt: { type: string, format: date-time, nullable: true }
    SearchTask:
      type: object
      required: [searchTaskId, status, sourceType, items, createdAt]
      properties:
        searchTaskId: { type: integer, format: int64 }
        status: { type: string, enum: [pending, succeeded, failed] }
        query: { type: string, nullable: true }
        sourceType:
          $ref: '#/components/schemas/SourceType'
        filters:
          type: object
          additionalProperties: true
        recognition:
          $ref: '#/components/schemas/Recognition'
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchTaskItem'
        suggestionCards:
          type: array
          items:
            $ref: '#/components/schemas/SuggestionCard'
        platformStats:
          type: array
          items:
            $ref: '#/components/schemas/PlatformStats'
        createdAt: { type: string, format: date-time }
    SearchTaskResponse:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SearchTask'
    RefineSearchTaskRequest:
      type: object
      required: [text]
      properties:
        text: { type: string }
        sortBy:
          $ref: '#/components/schemas/SortMode'
    RefineSearchTaskPayload:
      type: object
      required: [searchTaskId, text, filters, items, suggestionCards, platformStats, aiProvider, fallbackUsed, notices]
      properties:
        searchTaskId: { type: integer, format: int64 }
        text: { type: string }
        filters:
          type: object
          additionalProperties: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchTaskItem'
        suggestionCards:
          type: array
          items:
            $ref: '#/components/schemas/SuggestionCard'
        platformStats:
          type: array
          items:
            $ref: '#/components/schemas/PlatformStats'
        aiProvider:
          type: string
          example: rule
        fallbackUsed:
          type: boolean
        notices:
          type: array
          items: { type: string }
    RefineSearchTaskResponse:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RefineSearchTaskPayload'
    SearchTaskSummary:
      type: object
      required: [searchTaskId, status, sourceType, resultCount, createdAt]
      properties:
        searchTaskId: { type: integer, format: int64 }
        query: { type: string, nullable: true }
        status: { type: string, enum: [pending, succeeded, failed] }
        sourceType:
          $ref: '#/components/schemas/SourceType'
        resultCount: { type: integer }
        createdAt: { type: string, format: date-time }
    SearchTaskSummaryPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SearchTaskSummary'
    SearchTaskSummaryPageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SearchTaskSummaryPage'
...
    PageBase:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: {}
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }